# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TaskManager DOC API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Авторизация и Создание аккаунта"
                ],
                "summary": "Логин",
                "operationId": "ba24fa17938b90704dfbdb3b59591697",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "test@example.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "123123123"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "123"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 48000
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/users": {
            "post": {
                "tags": [
                    "Авторизация и Создание аккаунта"
                ],
                "summary": "Создание пользователя",
                "operationId": "ea53f04a382249f34755d11262e10e94",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Timmy"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "test@example.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "123123123"
                                            },
                                            "confirm_password": {
                                                "type": "string",
                                                "example": "123123123"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Timmy"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "test@example.com"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMTI3LjAuMC4xOjgwMDAvYXBpL3VzZXJzIiwiaWF0IjoxNzE5MjQzODAxLCJleHAiOjIwNzkyNDM4MDEsIm5iZiI6MTcxOTI0MzgwMSwianRpIjoidTRaeEdTVElEWDhyTlIyQiIsInN1YiI6IjE0IiwicHJ2IjoiMjNiZDVjODk0OWY2MDBhZGIzOWU3MDFjNDAwODcyZGI3YTU5NzZmNyJ9.mMgNmEDIP6xEnnIAO5WhrAXsW--uIHBInLbyCHmS9DY"
                                        },
                                        "profile": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Ivan"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "Dyrak"
                                                    },
                                                    "date_of_birth": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "10.10.1990"
                                                    },
                                                    "login": {
                                                        "type": "string",
                                                        "example": "Noone"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}